+++ index.js
@@ -1,2 +1,2 @@
 // The EasyRTC server files are in the lib folder
-module.exports = require('./lib/easyrtc_server');
+module.exports = require('./lib/easyrtc_public_obj');
