+++ lib/classes/connection.js
@@ -109,8 +109,14 @@ module.exports = function Connection(appName, socketId, easyrtcid, appObj, serve
      * @memberof    connectionObj
      * @return      {Object}    The application object
      */
-    connectionObj.getApp = function() {
-        return appObj;
+    connectionObj.getApp = function(callback) {
+
+        if (!callback) {
+            console.warn('use getApp with callback instead.');
+            return appObj;
+        }
+
+        return server.getAppName(appName, callback);
     };
