+++ lib/classes/event-emitter.js
@@ -14,7 +14,7 @@ const _ = require("underscore"); // General utility functions external module
 
 // Internals dependencies
 const easyrtc_utils = require("./../utils/error"); // General utility functions local module
-const eventListener = require("./../easyrtc_default_event_listeners"); // EasyRTC default event listeners
+const eventListener = require("./server/default-listeners"); // EasyRTC default event listeners
 
 // TODO migreate to ES6 or TypeScript
 module.exports = function EventEmitter(appName, socketId, easyrtcid, appObj) {
@@ -24,68 +24,6 @@ module.exports = function EventEmitter(appName, socketId, easyrtcid, appObj) {
 	 * @class
 	 */
 	const eventsObj = {};
-
-
-	/**
-	 * EasyRTC EventEmitter.
-	 *
-	 * @private
-	 */
-	eventsObj._eventListener = new events.EventEmitter();
-
-
-	/**
-	 * Expose event listener's emit function.
-	 *
-	 * @param       {string} eventName      EasyRTC event name.
-	 * @param       {...*} eventParam       The event parameters
-	 */
-	eventsObj.emit = eventsObj._eventListener.emit.bind(eventsObj._eventListener);
-
-
-	/**
-	 * Runs the default EasyRTC listener for a given event.
-	 *
-	 * @param       {string} eventName      EasyRTC event name.
-	 * @param       {...*} eventParam       The event parameters
-	 */
-	eventsObj.emitDefault = function() {
-	    if (!eventsObj.defaultListeners[arguments['0']]) {
-	        console.error("Error emitting listener. No default for event '" + arguments['0'] + "' exists.");
-	        return;
-	    }
-	    eventsObj.defaultListeners[Array.prototype.shift.call(arguments)].apply(this, arguments);
-	};
-
-
-	/**
-	 * Resets the listener for a given event to the default listener. Removes other listeners.
-	 *
-	 * @param       {string} eventName      EasyRTC event name.
-	 */
-	eventsObj.setDefaultListener = function(eventName) {
-	    if (!_.isFunction(eventsObj.defaultListeners[eventName])) {
-	        console.error("Error setting default listener. No default for event '" + eventName + "' exists.");
-	    }
-	    eventsObj._eventListener.removeAllListeners(eventName);
-	    eventsObj._eventListener.on(eventName, eventsObj.defaultListeners[eventName]);
-	};
-
-
-	/**
-	 * Resets the listener for all EasyRTC events to the default listeners. Removes all other listeners.
-	 */
-	eventsObj.setDefaultListeners = function() {
-	    eventsObj._eventListener.removeAllListeners();
-	    for (var currentEventName in eventsObj.defaultListeners) {
-	        if (_.isFunction(eventsObj.defaultListeners[currentEventName])) {
-	            eventsObj._eventListener.on(currentEventName, eventsObj.defaultListeners[currentEventName]);
-	        } else {
-	            throw new easyrtc_utils.ServerError("Error setting default listener. No default for event '" + currentEventName + "' exists.");
-	        }
-	    }
-	};
-
 
 	/**
 	 * Map of EasyRTC event listener names to their default functions. This map can be used to run a default function manually.
@@ -119,6 +57,63 @@ module.exports = function EventEmitter(appName, socketId, easyrtcid, appObj) {
 	    "msgTypeSetRoomApiField": eventListener.onMsgTypeSetRoomApiField
 	};
 
+	/**
+	 * EasyRTC EventEmitter.
+	 *
+	 * @private
+	 */
+	eventsObj._eventListener = new events.EventEmitter();
+
+	/**
+	 * Expose event listener's emit function.
+	 *
+	 * @param       {string} eventName      EasyRTC event name.
+	 * @param       {...*} eventParam       The event parameters
+	 */
+	eventsObj.emit = eventsObj._eventListener.emit.bind(eventsObj._eventListener);
+
+
+	/**
+	 * Runs the default EasyRTC listener for a given event.
+	 *
+	 * @param       {string} eventName      EasyRTC event name.
+	 * @param       {...*} eventParam       The event parameters
+	 */
+	eventsObj.emitDefault = function() {
+	    if (!eventsObj.defaultListeners[arguments['0']]) {
+	        console.error("Error emitting listener. No default for event '" + arguments['0'] + "' exists.");
+	        return;
+	    }
+	    eventsObj.defaultListeners[Array.prototype.shift.call(arguments)].apply(this, arguments);
+	};
+
+	/**
+	 * Resets the listener for a given event to the default listener. Removes other listeners.
+	 *
+	 * @param       {string} eventName      EasyRTC event name.
+	 */
+	eventsObj.setDefaultListener = function(eventName) {
+	    if (!_.isFunction(eventsObj.defaultListeners[eventName])) {
+	        console.error("Error setting default listener. No default for event '" + eventName + "' exists.");
+	    }
+	    eventsObj._eventListener.removeAllListeners(eventName);
+	    eventsObj._eventListener.on(eventName, eventsObj.defaultListeners[eventName]);
+	};
+
+
+	/**
+	 * Resets the listener for all EasyRTC events to the default listeners. Removes all other listeners.
+	 */
+	eventsObj.setDefaultListeners = function() {
+	    eventsObj._eventListener.removeAllListeners();
+	    for (var currentEventName in eventsObj.defaultListeners) {
+	        if (_.isFunction(eventsObj.defaultListeners[currentEventName])) {
+	            eventsObj._eventListener.on(currentEventName, eventsObj.defaultListeners[currentEventName]);
+	        } else {
+	            throw new easyrtc_utils.ServerError("Error setting default listener. No default for event '" + currentEventName + "' exists.");
+	        }
+	    }
+	};
 
 	/**
 	 * Sets listener for a given EasyRTC event. Only one listener is allowed per event. Any other listeners for an event are removed before adding the new one. See the events documentation for expected listener parameters.
@@ -135,7 +130,6 @@ module.exports = function EventEmitter(appName, socketId, easyrtcid, appObj) {
 	        throw new easyrtc_utils.ServerError("Unable to add listener to event '" + eventName + "'");
 	    }
 	};
-
 
 	/**
 	 * Removes all listeners for an event. If there is a default EasyRTC listener, it will be added. If eventName is `null`, all events will be removed than the defaults will be restored.
