+++ lib/classes/server.js
@@ -31,6 +31,8 @@ const Server = class Server {
     // TODO map readonly
     app = {};
 
+    socketServer = null;
+
     /**
      * EasyRTC Event handling object which contain most methods for interacting with EasyRTC events.
      * For convenience, this class has also been attached to the application, connection, session, and room classes.
@@ -72,7 +74,7 @@ const Server = class Server {
         }
     }
 
-    constructor() {
+    constructor(socketServer) {
         // Running the default listeners to initialize the events
         this.events.setDefaultListeners();
     }
