+++ lib/easyrtc_public_obj.js
@@ -21,6 +21,9 @@ const easyrtc_utils = require("./easyrtc_utils"); // General utility functions l
 const Server = require('./classes/server');
 const Application = require('./classes/application');
 
+// Check to ensure all required modules are available
+utils.checkModules();
+
 // Create new server isntance
 const easyrtcServer = new Server();
 const server = easyrtcServer;
@@ -69,18 +72,6 @@ pub.socketServer = null;
  * @member  {Object}    pub.httpApp
  */
 pub.httpApp = null;
-
-/**
- * Alias for easyrtc_utils object. Set during Listen()
- *
- * @member  {Object}    pub.util
- */
-pub.util = easyrtc_utils;
-
-/// TODO migrate to server
-pub.util.logCallback = (level, logText, logFields) => {
-    server.events.emit("log", level, logText, logFields);
-};
 
 /**
  * Returns a random available easyrtcid.
@@ -258,398 +249,28 @@ pub.listen = (httpApp, socketServer, options, listenCallback) => {
     // TODO listenCallback should return server
 }
 
-/**
- * Will attempt to deliver an EasyRTC session id via a cookie. Requires that session management be enabled from within Express.
- *
- * @param       {Object} req            Http request object
- * @param       {Object} res            Http result object
- */
-pub.util.sendSessionCookie = function(req, res) {
-    // If sessions or session cookies are disabled, return without an error.
-    if (!server.getOption("sessionEnable") || !server.getOption("sessionCookieEnable")) {
-        return;
-    }
+pub.shutdown = () => {
 
-    // TODO allow easyrtcsid value for cookie name options
-    if (req.sessionID && (!req.cookies || !req.cookies.easyrtcsid || req.cookieseasyrtcsid !== req.sessionID)) {
-        try {
-            easyrtc_utils.logDebug("Sending easyrtcsid cookie [" + req.sessionID + "] to [" + req.ip + "] for request [" + req.url + "]");
-            res.cookie("easyrtcsid", req.sessionID, {
-                maxAge: 2592000000,
-                httpOnly: false
-            });
-        } catch (e) {
-            easyrtc_utils.logWarning("Problem setting easyrtcsid cookie [" + req.sessionID + "] to [" + req.ip + "] for request [" + req.url + "]");
-        }
-    }
+    // Set server object references in public object
+    pub.httpApp         = null;
+    pub.socketServer    = null;
+    pub.server          = null;
+
 };
 
+//
+// TODO move util elsewhere
+//
+
 /**
- * Checks an incoming EasyRTC message to determine if it is syntactically valid.
+ * Alias for easyrtc_utils object. Set during Listen()
  *
- * @param       {string} type           The Socket.IO message type. Expected values are (easyrtcAuth|easyrtcCmd|easyrtcMsg)
- * @param       {Object} msg            Message object which contains the full message from a client; this can include the standard msgType and msgData fields.
- * @param       {?Object} appObj        EasyRTC application object. Contains methods used for identifying and managing an application.
- * @param       {function(?Error, boolean, string)} callback Callback with error, a boolean of whether message if valid, and a string indicating the error code if the message is invalid.
+ * @member  {Object}    pub.util
  */
-pub.util.isValidIncomingMessage = function(type, msg, appObj, callback) {
-    // A generic getOption variable which points to the getOption function at either the top or application level
-    var getOption = (_.isObject(appObj) ? appObj.getOption.bind(appObj) : server.getOption.bind(server));
+pub.util = easyrtc_utils;
 
-    // All messages follow the basic structure
-    if (!_.isString(type)) {
-        callback(null, false, "MSG_REJECT_BAD_TYPE");
-        return;
-    }
-    if (!_.isObject(msg)) {
-        callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-        return;
-    }
-    if (!_.isString(msg.msgType)) {
-        callback(null, false, "MSG_REJECT_BAD_TYPE");
-        return;
-    }
-
-    switch (type) {
-        case "easyrtcAuth":
-            if (msg.msgType !== "authenticate") {
-                callback(null, false, "MSG_REJECT_BAD_TYPE");
-                return;
-            }
-            if (!_.isObject(msg.msgData)) {
-                callback(null, false, "MSG_REJECT_BAD_DATA");
-                return;
-            }
-
-            // msgData.apiVersion (required)
-            if (msg.msgData.apiVersion === undefined || !_.isString(msg.msgData.apiVersion) || !getOption("apiVersionRegExp").test(msg.msgData.apiVersion)) {
-                callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                return;
-            }
-
-            // msgData.appName
-            if (msg.msgData.applicationName !== undefined && (!_.isString(msg.msgData.applicationName) || !getOption("appNameRegExp").test(msg.msgData.applicationName))) {
-                callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                return;
-            }
-
-            // msgData.easyrtcsid
-            if (msg.msgData.easyrtcsid !== undefined && (!_.isString(msg.msgData.easyrtcsid) || !getOption("easyrtcsidRegExp").test(msg.msgData.easyrtcsid))) {
-                callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                return;
-            }
-
-            var isCallbackRun = false;
-            async.waterfall([
-                function(asyncCallback) {
-                    if (!appObj) {
-                        pub.app((msg.msgData.applicationName !== undefined ? msg.msgData.applicationName : getOption("appDefaultName")), function(err, newAppObj) {
-                            if (!err) {
-                                appObj = newAppObj;
-                                getOption = appObj.getOption;
-                            }
-                            asyncCallback(null);
-                        });
-                    }
-                    else {
-                        asyncCallback(null);
-                    }
-                },
-                function(asyncCallback) {
-                    // msgData.username
-                    if (msg.msgData.username !== undefined && (!_.isString(msg.msgData.username) || !getOption("usernameRegExp").test(msg.msgData.username))) {
-                        callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                        isCallbackRun = true;
-                        return;
-                    }
-
-                    // msgData.credential
-                    if (msg.msgData.credential !== undefined && (!_.isObject(msg.msgData.credential) || _.isEmpty(msg.msgData.credential))) {
-                        callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                        isCallbackRun = true;
-                        return;
-                    }
-
-                    // msgData.roomJoin
-                    if (msg.msgData.roomJoin !== undefined) {
-                        if (!_.isObject(msg.msgData.roomJoin)) {
-                            callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                            isCallbackRun = true;
-                            return;
-                        }
-
-                        for (var currentRoomName in msg.msgData.roomJoin) {
-                            if (msg.msgData.roomJoin.hasOwnProperty(currentRoomName)) {
-                                if (
-                                    !getOption("roomNameRegExp").test(currentRoomName) ||
-                                        !_.isObject(msg.msgData.roomJoin[currentRoomName]) ||
-                                            !_.isString(msg.msgData.roomJoin[currentRoomName].roomName) ||
-                                                currentRoomName !== msg.msgData.roomJoin[currentRoomName].roomName
-                                ) {
-                                    callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                                    isCallbackRun = true;
-                                    return;
-                                }
-                                // if roomParameter field is defined, it must be an object
-                                if (msg.msgData.roomJoin[currentRoomName].roomParameter !== undefined && !_.isObject(msg.msgData.roomJoin[currentRoomName].roomParameter)) {
-                                    callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                                    isCallbackRun = true;
-                                    return;
-                                }
-                            }
-                        }
-                    }
-
-                    // msgData.setPresence
-                    if (msg.msgData.setPresence !== undefined) {
-                        if (!_.isObject(msg.msgData.setPresence) || _.isEmpty(msg.msgData.setPresence)) {
-                            callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                            isCallbackRun = true;
-                            return;
-                        }
-                        if (msg.msgData.setPresence.show !== undefined && (!_.isString(msg.msgData.setPresence.show) || !getOption("presenceShowRegExp").test(msg.msgData.setPresence.show))) {
-                            callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                            isCallbackRun = true;
-                            return;
-                        }
-                        if (msg.msgData.setPresence.status !== undefined && (!_.isString(msg.msgData.setPresence.status) || !getOption("presenceStatusRegExp").test(msg.msgData.setPresence.status))) {
-                            callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                            isCallbackRun = true;
-                            return;
-                        }
-                    }
-
-                    // TODO: setUserCfg
-                    if (msg.msgData.setUserCfg !== undefined) {
-
-                    }
-
-                    asyncCallback(null);
-
-                }
-            ],
-                function(err) {
-                    if (err) {
-                        if (!isCallbackRun) {
-                            callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                            isCallbackRun = true;
-                        }
-                    }
-                    else {
-                        // Incoming message syntactically valid
-                        callback(null, true, null);
-                    }
-                }
-            );
-
-            return; /* jshint ignore:line */
-        break; /* jshint ignore:line */
-
-        case "easyrtcCmd":
-            switch (msg.msgType) {
-                case "candidate" :
-                case "offer" :
-                case "answer" :
-                    // candidate, offer, and answer each require a non-empty msgData object and a proper targetEasyrtcid
-                    if (!_.isObject(msg.msgData) || _.isEmpty(msg.msgData)) {
-                        callback(null, false, "MSG_REJECT_BAD_DATA");
-                        return;
-                    }
-                    if (!_.isString(msg.targetEasyrtcid) || !getOption("easyrtcidRegExp").test(msg.targetEasyrtcid)) {
-                        callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                        return;
-                    }
-                    break;
-                case "reject" :
-                case "hangup" :
-                    // reject, and hangup each require a targetEasyrtcid but no msgData
-                    if (msg.msgData !== undefined) {
-                        callback(null, false, "MSG_REJECT_BAD_DATA");
-                        return;
-                    }
-                    if (!_.isString(msg.targetEasyrtcid) || !getOption("easyrtcidRegExp").test(msg.targetEasyrtcid)) {
-                        callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                        return;
-                    }
-                    break;
-
-                case "getIceConfig" :
-                    if (msg.msgData !== undefined && !_.isEmpty(msg.msgData)) {
-                        callback(null, false, "MSG_REJECT_BAD_DATA");
-                        return;
-                    }
-                    break;
-
-                case "getRoomList" :
-                    if (msg.msgData !== undefined) {
-                        callback(null, false, "MSG_REJECT_BAD_DATA");
-                        return;
-                    }
-                    break;
-
-                case "roomJoin" :
-                    if (!_.isObject(msg.msgData)) {
-                        callback(null, false, "MSG_REJECT_BAD_DATA");
-                        return;
-                    }
-                    if (!_.isObject(msg.msgData.roomJoin)) {
-                        callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                        return;
-                    }
-
-                    for (var joinRoomName in msg.msgData.roomJoin) {
-                        if (msg.msgData.roomJoin.hasOwnProperty(joinRoomName)) {
-                            if (
-                                !getOption("roomNameRegExp").test(joinRoomName) ||
-                                    !_.isObject(msg.msgData.roomJoin[joinRoomName]) ||
-                                        !_.isString(msg.msgData.roomJoin[joinRoomName].roomName) ||
-                                            joinRoomName !== msg.msgData.roomJoin[joinRoomName].roomName
-                            ) {
-                                callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                                return;
-                            }
-                        }
-                    }
-                    break;
-
-                case "roomLeave" :
-                    if (!_.isObject(msg.msgData)) {
-                        callback(null, false, "MSG_REJECT_BAD_DATA");
-                        return;
-                    }
-                    if (!_.isObject(msg.msgData.roomLeave)) {
-                        callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                        return;
-                    }
-
-                    for (var leaveRoomName in msg.msgData.roomLeave) {
-                        if (msg.msgData.roomLeave.hasOwnProperty(leaveRoomName)) {
-                            if (
-                                !getOption("roomNameRegExp").test(leaveRoomName) ||
-                                    !_.isObject(msg.msgData.roomLeave[leaveRoomName]) ||
-                                        !_.isString(msg.msgData.roomLeave[leaveRoomName].roomName) ||
-                                            leaveRoomName !== msg.msgData.roomLeave[leaveRoomName].roomName) {
-                                callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                                return;
-                            }
-                        }
-                    }
-                    break;
-
-                case "stillAlive":
-                    // Deprecated
-                    break;
-
-                case "setPresence" :
-                    if (!_.isObject(msg.msgData)) {
-                        callback(null, false, "MSG_REJECT_BAD_DATA");
-                        return;
-                    }
-                    if (!_.isObject(msg.msgData.setPresence) || _.isEmpty(msg.msgData.setPresence)) {
-                        callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                        return;
-                    }
-                    if (msg.msgData.setPresence.show !== undefined && (!_.isString(msg.msgData.setPresence.show) || !getOption("presenceShowRegExp").test(msg.msgData.setPresence.show))) {
-                        callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                        return;
-                    }
-                    if (msg.msgData.setPresence.status !== undefined && (!_.isString(msg.msgData.setPresence.status) || !getOption("presenceStatusRegExp").test(msg.msgData.setPresence.status))) {
-                        callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                        return;
-                    }
-                    break;
-
-                case "setRoomApiField" :
-                    if (!_.isObject(msg.msgData)) {
-                        callback(null, false, "MSG_REJECT_BAD_DATA");
-                        return;
-                    }
-                    if (!_.isObject(msg.msgData.setRoomApiField) || _.isEmpty(msg.msgData.setRoomApiField)) {
-                        callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                        return;
-                    }
-                    if (!_.isString(msg.msgData.setRoomApiField.roomName) || !getOption("roomNameRegExp").test(msg.msgData.setRoomApiField.roomName)) {
-                        callback(null, false, "MSG_REJECT_BAD_ROOM");
-                        return;
-                    }
-                    if (msg.msgData.setRoomApiField.field !== undefined) {
-                        if (!_.isObject(msg.msgData.setRoomApiField.field)) {
-                            callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                            return;
-                        }
-                        try {
-                            if (JSON.stringify(msg.msgData.setRoomApiField.field).length >= 4096) {
-                                callback(null, false, "MSG_REJECT_BAD_SIZE");
-                                return;
-                            }
-                        } catch (e) {
-                            if (!_.isObject(msg.msgData.setRoomApiField.field)) {
-                                callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                                return;
-                            }
-                        }
-                    }
-                    break;
-
-                case "setUserCfg" :
-                    if (!_.isObject(msg.msgData)) {
-                        callback(null, false, "MSG_REJECT_BAD_DATA");
-                        return;
-                    }
-                    if (!_.isObject(msg.msgData.setUserCfg) || _.isEmpty(msg.msgData.setUserCfg)) {
-                        callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                        return;
-                    }
-
-                    // setUserCfg.p2pList
-                    if (msg.msgData.setUserCfg.p2pList !== undefined && (!_.isObject(msg.msgData.setUserCfg.p2pList) || _.isEmpty(msg.msgData.setUserCfg.p2pList))) {
-                        callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                        return;
-                    }
-                    // TODO: Go through p2pList to confirm each key is an easyrtcid
-
-                    // setUserCfg.userSettings
-                    if (msg.msgData.setUserCfg.userSettings !== undefined && (!_.isObject(msg.msgData.setUserCfg.userSettings) || _.isEmpty(msg.msgData.setUserCfg.userSettings))) {
-                        callback(null, false, "MSG_REJECT_BAD_STRUCTURE");
-                        return;
-                    }
-
-                    break;
-
-                default:
-                    // Reject all unknown msgType's
-                    callback(null, false, "MSG_REJECT_BAD_TYPE");
-                    return;
-            }
-
-            break;
-
-        case "easyrtcMsg":
-            // targetEasyrtcid
-            if (msg.targetEasyrtcid !== undefined && (!_.isString(msg.targetEasyrtcid) || !getOption("easyrtcidRegExp").test(msg.targetEasyrtcid))) {
-                callback(null, false, "MSG_REJECT_TARGET_EASYRTCID");
-                return;
-            }
-            // targetGroup
-            if (msg.targetGroup !== undefined && (!_.isString(msg.targetGroup) || !getOption("groupNameRegExp").test(msg.targetGroup))) {
-                callback(null, false, "MSG_REJECT_TARGET_GROUP");
-                return;
-            }
-            // targetRoom
-            if (msg.targetRoom !== undefined && (!_.isString(msg.targetRoom) || !getOption("roomNameRegExp").test(msg.targetRoom))) {
-                callback(null, false, "MSG_REJECT_TARGET_ROOM");
-                return;
-            }
-            break;
-
-        default:
-            callback(null, false, "MSG_REJECT_BAD_TYPE");
-            return;
-    }
-
-    // Incoming message syntactically valid
-    callback(null, true, null);
-};
+/// TODO migrate to server
+pub.util.logCallback = (level, logText, logFields) => server.events.emit("log", level, logText, logFields);
 
 // Documenting global callbacks
