+++ lib/easyrtc_server.js
@@ -1,26 +1,16 @@
-/* global module, require, process */
+/* global module, require */
 
 /**
- * @file        Entry library for EasyRTC server. Houses the primary listen function.
+ * @file        Default options used within EasyRTC. Overriding of default options should be done using the public listen() or setOption() functions.
+ * @module      easyrtc_server
  * @author      Open-EasyRTC
  * @copyright   Copyright 2022 Open-EasyRTC. All rights reserved.
  * @license     BSD v2, see LICENSE file in module root folder.
  */
 
-// Externals dependencies
-var _ = require("underscore"); // General utility functions external module
+// EasyRTC global variable
+const defaultOptions = require("./easyrtc_public_obj");
+console.warn("Deprecated use '/easyrtc_public_obj' instead");
 
-// Internals dependencies
-var server      = require("./easyrtc_public_obj"); // EasyRTC public object
-var utils    = require("./utils"); // General utility functions local module
-
-utils.checkModules(); // Check to ensure all required modules are available
-
-/**
- * Gets EasyRTC Version. The format is in a major.minor.patch format with an optional letter following denoting alpha or beta status. The version is retrieved from the package.json file located in the EasyRTC project root folder.
- *
- * @return      {string}                EasyRTC Version
- */
-server.getVersion = () => utils.getPackageData("version");
-
-module.exports = server;
+// Allows the option object to be seen by the caller.
+module.exports = defaultOptions;
