+++ lib/utils/error.js
@@ -1,3 +1,13 @@
+/* global module, require */
+
+/**
+ * @file        Maintains private object used within EasyRTC for holding in-memory state information
+ * @module      OpenEasyRTC
+ * @author      Open-EasyRTC
+ * @copyright   Copyright 2022 Open-EasyRTC. All rights reserved.
+ * @license     BSD v2, see LICENSE file in module root folder.
+ */
+
 const util = require("util");
 
 const easyrtc_error = module.exports;
