+++ lib/utils/log.js
@@ -1,4 +1,12 @@
+/* global module, require */
 
+/**
+ * @file        Maintains private object used within EasyRTC for holding in-memory state information
+ * @module      OpenEasyRTC
+ * @author      Open-EasyRTC
+ * @copyright   Copyright 2022 Open-EasyRTC. All rights reserved.
+ * @license     BSD v2, see LICENSE file in module root folder.
+ */
 
 var easyrtc_log = module.exports = {
 
@@ -42,7 +50,6 @@ easyrtc_log.log = function(level, logText, logFields, callback) {
     }
 };
 
-
 /**
  * Convenience function for logging "debug" level items.
  *
@@ -52,7 +59,6 @@ easyrtc_log.log = function(level, logText, logFields, callback) {
 easyrtc_log.logDebug = function(logText, logFields) {
     easyrtc_log.log("debug", logText, logFields);
 };
-
 
 /**
  * Convenience function for logging "info" level items.
@@ -64,7 +70,6 @@ easyrtc_log.logInfo = function(logText, logFields) {
     easyrtc_log.log("info", logText, logFields);
 };
 
-
 /**
  * Convenience function for logging "warning" level items.
  *
@@ -74,7 +79,6 @@ easyrtc_log.logInfo = function(logText, logFields) {
 easyrtc_log.logWarning = function(logText, logFields) {
     easyrtc_log.log("warning", logText, logFields);
 };
-
 
 /**
  * Convenience function for logging "error" level items.
